メール自動送信までの修正
