+++ app/models/User/User.rb
@@ -9,15 +9,16 @@
 #  created_at      :datetime         not null
 #  updated_at      :datetime         not null
 #
+module User
+  class User < ApplicationRecord
+    has_secure_password
+    has_one :user_profile, dependent: :destroy
+    has_one :provisional_user_completed_log, dependent: :destroy
+    has_many :user_freezed_reasons, dependent: :destroy
+    has_many :user_unfreezed_reasons, dependent: :destroy
+    has_many :user_auth_logs, dependent: :destroy
 
-class User < ApplicationRecord
-  has_secure_password
-  has_one :user_profile, dependent: :destroy
-  has_one :provisional_user_completed_log, dependent: :destroy
-  has_many :user_freezed_reasons, dependent: :destroy
-  has_many :user_unfreezed_reasons, dependent: :destroy
-  has_many :user_auth_logs, dependent: :destroy
-
-  validates :email, presence: true
-  validates :password_digest, presence: true
+    validates :email, presence: true
+    validates :password_digest, presence: true
+  end
 end
