+++ app/models/User/user_profile.rb
@@ -13,13 +13,14 @@
 #  created_at      :datetime         not null
 #  updated_at      :datetime         not null
 #
-
-class UserProfile < ApplicationRecord
-  belongs_to :user
-  validates :last_name, presence: true, length: { maximum: 20 }
-  validates :last_name_kana, presence: true, length: { maximum: 30 }
-  validates :first_name, presence: true, length: { maximum: 20 }
-  validates :first_name_kana, presence: true, length: { maximum: 30 }
-  validates :birth_on, presence: true
-  validates :job, presence: true
+module User
+  class UserProfile < ApplicationRecord
+    belongs_to :user
+    validates :last_name, presence: true, length: { maximum: 20 }
+    validates :last_name_kana, presence: true, length: { maximum: 30 }
+    validates :first_name, presence: true, length: { maximum: 20 }
+    validates :first_name_kana, presence: true, length: { maximum: 30 }
+    validates :birth_on, presence: true
+    validates :job, presence: true
+  end
 end
