--- app/views/layouts/application.html.erb
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Rediction</title>
-    <%= csrf_meta_tags %>
-    <%= csp_meta_tag %>
-
-    <%= stylesheet_link_tag    'application', media: 'all', 'data-turbolinks-track': 'reload' %>
-    <%= javascript_include_tag 'application', 'data-turbolinks-track': 'reload' %>
-  </head>
-
-  <body>
-    <% if flash[:success] %>
-       <p><%= flash[:success]%></p>
-    <% end %>
-    <%= yield %>
-  </body>
-</html>
