+++ app/views/layouts/application.html.slim
@@ -1,0 +1,14 @@
+doctype html
+html
+  head
+    title Rediction
+    = csrf_meta_tags
+    = csp_meta_tag
+
+    = stylesheet_link_tag    'application', media: 'all', 'data-turbolinks-track': 'reload'
+    = javascript_include_tag 'application', 'data-turbolinks-track': 'reload'
+  body
+  - flash.each do |type, message|
+      #flash-msg.flash class="flash-#{type}"
+        = message
+  = yield
