+++ app/views/layouts/mailer.html.slim
@@ -1,0 +1,7 @@
+doctype html
+html
+  head
+    meta http-equiv="Content-Type" content="text/html; charset=utf-8"
+    style
+  body
+    = yield
