+++ app/views/layouts/mailer.text.slim
@@ -1,0 +1,1 @@
+= yield
