+++ app/views/provisional_users/create.html.slim
@@ -1,0 +1,2 @@
+h1
+  | メールが送信されましたので、ご確認ください。
