+++ app/views/registeration_mailer/send_when_registeration.text.slim
@@ -1,9 +1,8 @@
 
 ようこそredictionへ
 
-===============================================
 
-このサイトにログインするには、
+この
 = @url
-をクリックしてください。
+をクリックして会員登録を完了させてください
