+++ app/views/root/index.html.slim
@@ -1,7 +1,4 @@
 h1
-  | Hello World / Tere, Maailm / Привет мир
-
-p
-  | ちゃっちゃと最高なもの作ろうぜ！！
+  | トップページ
 br
-= link_to("さぁまずは会員登録だ！", new_user_path)
+= link_to "さぁまずは会員登録だ！", new_provisional_user_path
