+++ config/locales/en.yml
@@ -31,3 +31,4 @@
 
 en:
   hello: "Hello world"
+
