+++ config/locales/ja.yml
@@ -1,2 +1,6 @@
 ja:
   hello: 'こんにちは'
+  last_name: '苗字'
+  last_name_kana: 'フリガナ(苗字)'
+  first_name: '名前'
+  first_name_kana: 'フリガナ(名前)'
